backend/pyproject.toml
[project]
name = "ai-news-backend"
version = "0.1.0"
description = "AI News Backend with OpenAI Agents SDK"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "openai>=1.54.0",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "reportlab>=4.0.0",
    "matplotlib>=3.8.0",
    "gtts>=2.5.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "utils"]

